requirements.txt
scipy
keras
tensorflow
sklearn
scikit-image
matplotlib
jupyter
pandas
texttable

Cython
python-rtmidi
mido
mypy


pyknon
